Refine/lovesick automember (#5)
* feat(util): Add date util

* feat(lovesick-automember): Change strategy to check by start of month, refactor some code

* docs(lovesick): Fix typo

* feat(util): Add more date util

* feat(lovesick-automember): Add backup after check

* feat(lovesick-automember): Set next check time to start of day

* fix(lovesick): Remove trailing comma in message template

* feat(paginator): Add DataEmbedSource

* refactor(lovesick): Refactor inactive member tracking

This commit refactors the way inactive member data is tracked in the LoveSick extension. Instead of using separate lists for member IDs and member strings, the data is now stored in a single list of dictionaries. Each dictionary contains the member ID, name, and the number of counts. This change improves the efficiency and readability of the code.

Refactor the inactive member tracking in LoveSick extension to use a list of dictionaries instead of separate lists for member IDs and member strings.